requirements.txt
networkx==2.3
numpy==1.17.2
tqdm==4.36.1
chardet==3.0.4
lxml==4.4.1
scipy==1.3.1
seaborn==0.9.0
matplotlib==3.1.3
graphviz==0.13.2
treelib==1.5.5
Pillow==8.2.0
heurfh
